updated StoreFactory to add ArrayStorePartition factory methods

+++ src/test/java/test/util/FileUtils.java
@@ -43,33 +43,35 @@ public class FileUtils {
     
     public static void cleanDirectory(File dir) throws IOException {
         File[] files = dir.listFiles();
-
-        for (File f : files) {
-            if (f.isFile()) {
-                boolean deleted = f.delete();
-                if (!deleted) {
-                    throw new IOException("file:" + f.getAbsolutePath() + " not deleted");
+        if(files != null) {
+            for (File f : files) {
+                if (f.isFile()) {
+                    boolean deleted = f.delete();
+                    if (!deleted) {
+                        throw new IOException("file:" + f.getAbsolutePath() + " not deleted");
+                    }
+                } else {
+                    deleteDirectory(f);
                 }
-            } else {
-                deleteDirectory(f);
             }
         }
     }
 
     public static void deleteDirectory(File dir) throws IOException {
         File[] files = dir.listFiles();
-
-        for (File f : files) {
-            if (f.isDirectory()) {
-                deleteDirectory(f);
-            } else {
-                boolean deleted = f.delete();
-                if (!deleted) {
-                    throw new IOException("file:" + f.getAbsolutePath() + " not deleted");
+        if(files != null) {
+            for (File f : files) {
+                if (f.isDirectory()) {
+                    deleteDirectory(f);
+                } else {
+                    boolean deleted = f.delete();
+                    if (!deleted) {
+                        throw new IOException("file:" + f.getAbsolutePath() + " not deleted");
+                    }
                 }
             }
         }
-
+        
         boolean deleted = dir.delete();
         if (!deleted) {
             throw new IOException("dir:" + dir.getAbsolutePath() + " not deleted");
